refactor(retrieval-api): extract page-context matching to api/page_context.py
Behavior-preserving lift of the page-context URL cluster
(_normalise_page_context_url, _same_page_context_path, _apply_page_context_boost
+ the _PAGE_CONTEXT_SCORE_BOOST tuning constant) out of retrieve.py. The two URL
helpers and the constant are internal to the new module; retrieve re-imports only
_apply_page_context_boost so the orchestrator call sites and the
`from retrieval_api.api.retrieve import _apply_page_context_boost` test import are
unchanged. Dropped the now-unused urllib.parse import from retrieve.

retrieval-api suite: 530 passed, 1 skipped. ruff clean.
