+++ klai-retrieval-api/retrieval_api/api/retrieve.py
@@ -8,12 +8,12 @@ import math
 import os
 import time
 import uuid
-from urllib.parse import urlparse, urlunparse
 
 import structlog
 from fastapi import APIRouter, Depends, HTTPException, Request
 from klai_kb_slugs import personal_kb_slug
 
+from retrieval_api.api.page_context import _apply_page_context_boost
 from retrieval_api.api.ranking import (
     _apply_link_expand_boost,
     _compute_confidence_band,
@@ -63,7 +63,6 @@ _RETRIEVAL_QUERY_SCOPE = "klai:internal:retrieval:query"
 # Module-level singleton — avoids ruff B008 ("Depends in default arg") and
 # is the FastAPI-recommended pattern for repeated dependencies.
 _REQUIRE_RETRIEVAL_SCOPE = Depends(require_scope(_RETRIEVAL_QUERY_SCOPE))
-_PAGE_CONTEXT_SCORE_BOOST = 1.08
 
 
 def _caller_pre_resolved(req: RetrieveRequest) -> bool:
@@ -76,69 +75,6 @@ def _caller_pre_resolved(req: RetrieveRequest) -> bool:
     both fall through to retrieval-side coreference resolution.
     """
     return bool(req.raw_query) and req.raw_query != req.query
-
-
-def _normalise_page_context_url(raw_url: str | None) -> str:
-    if not raw_url:
-        return ""
-    try:
-        parsed = urlparse(raw_url.strip())
-    except ValueError:
-        return ""
-    if parsed.scheme not in {"http", "https"} or not parsed.netloc:
-        return ""
-    path = parsed.path.rstrip("/") or "/"
-    return urlunparse((parsed.scheme.lower(), parsed.netloc.lower(), path, "", "", ""))
-
-
-def _same_page_context_path(source_url: str, page_url: str) -> bool:
-    source = urlparse(source_url)
-    page = urlparse(page_url)
-    if source.scheme != page.scheme or source.netloc != page.netloc:
-        return False
-    source_path = source.path.rstrip("/") or "/"
-    page_path = page.path.rstrip("/") or "/"
-    if source_path == "/" or page_path == "/":
-        return source_path == page_path
-    return source_path.startswith(f"{page_path}/") or page_path.startswith(f"{source_path}/")
-
-
-def _apply_page_context_boost(
-    chunks: list[dict],
-    page_context: dict[str, str] | None,
-    *,
-    mark: bool = True,
-) -> tuple[list[dict], int]:
-    page_url = _normalise_page_context_url((page_context or {}).get("url"))
-    if not page_url:
-        return chunks, 0
-
-    boosted_count = 0
-    for chunk in chunks:
-        source_url = _normalise_page_context_url(chunk.get("source_url"))
-        if not source_url:
-            continue
-        if source_url != page_url and not _same_page_context_path(source_url, page_url):
-            continue
-
-        boosted_count += 1
-        score_key = (
-            "reranker_score" if isinstance(chunk.get("reranker_score"), (int, float)) else "score"
-        )
-        if isinstance(chunk.get(score_key), (int, float)):
-            boosted_score = chunk[score_key] * _PAGE_CONTEXT_SCORE_BOOST
-            chunk[score_key] = (
-                min(boosted_score, 1.0) if score_key == "reranker_score" else boosted_score
-            )
-            if mark:
-                chunk["_page_context_boosted"] = True
-
-    if boosted_count:
-        chunks.sort(
-            key=lambda c: c.get("reranker_score") or c.get("score") or 0.0,
-            reverse=True,
-        )
-    return chunks, boosted_count
 
 
 def _evidence_pack_decision_sources(evidence_pack: object) -> list[dict[str, object]]:
